feat(app): Enhance install to merge with exiting installation
When a user is doing another installation, he may just choose the new
Qt directory (not re-install into old one).
We merge this directory with the one he installed too, so they will all
be updated.

IMO it's closer to what is expected: if installed once, the user expects
to have it updated later on.