test: Add a test project for natvis files
This project should be open in Visual Studio and the user should be able
to run the project and see the natvis files in action.
Note that we can't support VS Code yet, as it's not able to load
multiple natvis files at once.
See: https://github.com/microsoft/vscode-cpptools/issues/10917